store/postgres: Move `check_index_is_valid` to the `catalog` module
