+++ store/postgres/src/deployment.rs
@@ -11,7 +11,7 @@ use diesel::{expression::SqlLiteral, pg::PgConnection, sql_types::Numeric};
 use diesel::{
     prelude::{ExpressionMethods, OptionalExtension, QueryDsl, RunQueryDsl},
     sql_query,
-    sql_types::{Bool, Nullable, Text},
+    sql_types::{Nullable, Text},
 };
 use graph::data::subgraph::{schema::SubgraphManifestEntity, SubgraphFeature};
 use graph::prelude::{
@@ -896,37 +896,4 @@ pub fn set_entity_count(
         .set(d::entity_count.eq(sql(&full_count_query)))
         .execute(conn)?;
     Ok(())
-}
-
-/// Helper type for checking if an index was created successfully.
-#[derive(Queryable, QueryableByName)]
-struct ManualIndexCheck {
-    #[sql_type = "Bool"]
-    is_valid: bool,
-}
-
-/// Checks in the database if a given index is valid.
-pub(crate) fn check_index_is_valid(
-    conn: &PgConnection,
-    schema_name: &str,
-    index_name: &str,
-) -> Result<bool, StoreError> {
-    let query = "
-        select
-            i.indisvalid as is_valid
-        from
-            pg_class c
-            join pg_index i on i.indexrelid = c.oid
-            join pg_namespace n on c.relnamespace = n.oid
-        where
-            n.nspname = $1
-            and c.relname = $2";
-    let result = sql_query(query)
-        .bind::<Text, _>(schema_name)
-        .bind::<Text, _>(index_name)
-        .get_result::<ManualIndexCheck>(conn)
-        .optional()
-        .map_err::<StoreError, _>(Into::into)?
-        .map(|check| check.is_valid);
-    Ok(matches!(result, Some(true)))
 }
